Cleanup `@internal` APIs from `presentation-testing` barrel
