+++ packages/testing/package.json
@@ -35,7 +35,7 @@
     "test": "mocha --config ./.mocharc.json \"./lib/cjs/test/**/*.test.js\"",
     "test:regression": "npm -s test",
     "extract-api": "betools extract-api --entry=presentation-testing --apiReportFolder=./api --apiReportTempFolder=./api/temp --apiSummaryFolder=./api",
-    "check-internal": ""
+    "check-internal": "node ../../scripts/checkInternal.js --apiSummary ./api/presentation-testing.api.md"
   },
   "dependencies": {
     "@itwin/presentation-components": "workspace:*",
