+++ packages/testing/src/presentation-testing/IModelBuilderImpl.ts
@@ -1,0 +1,49 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+/** @packageDocumentation
+ * @module IModel
+ */
+
+import { IModelDb } from "@itwin/core-backend";
+import { Id64String } from "@itwin/core-bentley";
+import { BisCodeSpec, Code, CodeScopeProps, CodeSpec, ElementAspectProps, ElementProps, ModelProps, RelationshipProps } from "@itwin/core-common";
+import { TestIModelBuilder } from "./IModelUtilities";
+
+/**
+ * Default implementation of the IModel builder interface.
+ * @internal
+ */
+export class TestIModelBuilderImpl implements TestIModelBuilder {
+  private _iModel: IModelDb;
+
+  constructor(iModel: IModelDb) {
+    this._iModel = iModel;
+  }
+
+  public insertModel<TProps extends ModelProps>(props: TProps): Id64String {
+    return this._iModel.models.insertModel(props);
+  }
+
+  public insertElement<TProps extends ElementProps>(props: TProps): Id64String {
+    return this._iModel.elements.insertElement(props);
+  }
+
+  public insertAspect<TProps extends ElementAspectProps>(props: TProps): Id64String {
+    return this._iModel.elements.insertAspect(props);
+  }
+
+  public insertRelationship<TProps extends RelationshipProps>(props: TProps): Id64String {
+    return this._iModel.relationships.insertInstance(props);
+  }
+
+  public createCode(scopeModelId: CodeScopeProps, codeSpecName: BisCodeSpec, codeValue: string): Code {
+    const codeSpec: CodeSpec = this._iModel.codeSpecs.getByName(codeSpecName);
+    return new Code({ spec: codeSpec.id, scope: scopeModelId, value: codeValue });
+  }
+
+  public async importSchema(schemaXml: string) {
+    await this._iModel.importSchemaStrings([schemaXml]);
+  }
+}
